% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make2001NHTSDataset.r
\name{getZipDatasetFromRepo}
\alias{getZipDatasetFromRepo}
\title{Retrieve data in zip archive from repository}
\usage{
getZipDatasetFromRepo(Repo, DatasetName)
}
\arguments{
\item{Repo}{A string that is the url where the zip archive is located.}

\item{DatasetName}{The name of the dataset.}
}
\value{
A data frame containing the data in the zip archive.
}
\description{
\code{getZipDatasetFromRepo} retrieve a zip archive containing a csv file
from a repository, unzip, and read as data frame
}
\details{
This function retrieves a zip archive containing a csv file from a
repository, unzips it, and reads it in as data frame which it returns.
}
