% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make2001NHTSDataset.r
\docType{data}
\name{Per_df}
\alias{Per_df}
\title{Person dataset from the 2001 National Household Travel Survey}
\format{An object of class \code{data.frame} with 144884 rows and 15 columns.}
\usage{
Per_df
}
\description{
A dataset of person characteristics derived from the 2001 National Household
Travel Survey and used in the estimation of several VisionEval models.
}
\details{
@format A data frame with 144884 rows and 14 columns.
 \describe{
   \item{Houseid}{Unique household ID}
   \item{Personid}{Unique ID of person in household}
   \item{Commdrvr}{Person is a commercial driver}
   \item{Nbiketrp}{Number of bike trips in past week}
   \item{Nwalktrp}{Number of walk trips in past week}
   \item{Usepubtr}{Used public transit on travel day}
   \item{Wrkdrive}{Job requires driving a motor vehicle}
   \item{Wrktrans}{Transportation mode to work last week}
   \item{Worker}{Person has job (1=yes, 0=no)}
   \item{Dtgas}{Price of gasoline}
   \item{Disttowk}{Distance to work in miles}
   \item{Driver}{Driver status (1=driver, 0=non-driver)}
   \item{R_age}{Age}
   \item{R_sex}{Sex}
   \item{AgeGroup}{AgeGroup: Age0to14, Age15to19, Age20to29, Age30to54, Age55to64, Age65Plus}
 }
 @source 2001 National Household Travel Survey and Make2001NHTSDataset.R script.
}
\keyword{datasets}
