% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make2001NHTSDataset.r
\docType{data}
\name{Veh_df}
\alias{Veh_df}
\title{Vehicle dataset from the 2001 National Household Travel Survey}
\format{A data frame with 112697 rows and 10 variables
\describe{
  \item{Houseid}{Unique household ID}
  \item{Vehid}{Unique ID for vehicle in household}
  \item{Bestmile}{Best estimate of annual miles}
  \item{Eiadmpg}{EIA derived miles per equivalent-gallon}
  \item{Gscost}{Estimated Fuel cost (cents per gallon)}
  \item{Vehtype}{Type of vehicle}
  \item{Vehyear}{Vehicle year - derived}
  \item{Vehmiles}{Miles vehicle driven last 12 months}
  \item{Type}{Auto or light truck}
  \item{Gscostmile2}{Estimated gas cost per mile of travel}
}}
\source{
2001 National Household Travel Survey and Make2001NHTSDataset.R script.
}
\usage{
Veh_df
}
\description{
A vehicle dataset containing the data used for estimating VisionEval
vehicle models derived from the 2001 National Household Travel Survey.
}
\keyword{datasets}
